+++ shiny/api-examples/data_frame/app-express.py
@@ -66,8 +66,9 @@ with ui.layout_columns(col_widths=[12, 6, 6]):
 
 @reactive.calc
 def filtered_df():
-    req(not summary_data.data_selected().empty)
-    countries = summary_data.data_selected()["country"]
+    data_selected = summary_data.data_view(selected=True)
+    req(not data_selected.empty)
+    countries = data_selected["country"]
 
     # Filter data for selected countries
     return df[df["country"].isin(countries)]