+++ shiny/render/_data_frame_utils/__init__.py
@@ -18,11 +18,10 @@ from ._patch import (
     cell_patch_processed_to_jsonifiable,
 )
 from ._selection import (
-    BrowserCellSelection,
     CellSelection,
     SelectionMode,
     SelectionModes,
-    as_browser_cell_selection,
+    as_cell_selection,
 )
 
 __all__ = (
@@ -41,9 +40,8 @@ __all__ = (
     "PatchFnSync",
     "assert_patches_shape",
     "cell_patch_processed_to_jsonifiable",
-    "BrowserCellSelection",
     "CellSelection",
     "SelectionMode",
     "SelectionModes",
-    "as_browser_cell_selection",
+    "as_cell_selection",
 )